perf(kubernetes): Improve performance of cache kind lookup
Looking up a kind and a logicalKind by name (which happens frequently
on reads from the cache) are fairly inefficient.  We use a stream
to iterate over all current enum values comparing the input string to
calling toString on the enum using a case-insensitive comparison.

toString is unnecessarily expensive as it allocates a new string on
each call via calling .toLowerCase.  Instead, store the lower-case
name as a property of the enum and return a reference to it in toString.

Improve fromString by just converting the input string to upper case
and using the enum's valueOf function, rather than looping over
the values and doing a case-insensitive search.