test(kubernetes): Add tests to Keys
An upcoming commit will make some performance improvements to the
Keys.Kind and Keys.LogicalKind enums, so add some tests to these.

Also add some tests to validate the parsing of cache keys containing
semicolons. This is currently broken, but add the test now validating
the incorrect behavior and update the test in the commit that fixes it.