fix(kubernetes): Improve failure mode for unreachable cluster
We currently cache any call to get the namespaces for an account
with an expiry time of 30 s using a memoized supplier.

When a cluster is unreachable, the call to get the cluster's
namespaces will hang and eventually time out; we then log a
warning and return an empty array of namesapces.

If the call to kubectl returns an error, we don't cache the empty
list return value, so every call to get namespaces will call
kubectl.  This leads to a bad failure mode where a slow/unresponsive
cluster leads to more calls than a fast/responsive cluster.

To address this, when a call to get namespaces returns an error,
cache the empty list we're returning for the same amount of time
as a successful call.