fix(edda): Better error message on edda lookup failures
Previously, when edda lookup for a loadbalancer/target group would fail we would get
a useless stack trace with `UndeclaredThrowableException` at the top. This exception
would show up in deck causing lots of user confusion (and fail execution).
When in fact, such errors are often benign (we already ignore similar failures of direct AWS lookups)

This change plumbs through the 404 from edda so that it can be dealt with appropriately
(all other failure remain unchanged and re-throw the exception)