perf(kubernetes): Use a more efficient data structure for cache data
The current logic to build the Kubernetes cache data first creates
a list of CacheData objects representing each of the objects to be
cached (including its attributes and relationships).

At the end of the caching process, we want only a single CacheData
object for each cache key and also all relationships to be
bidirectional.

Bidirectionality is currently enforced by calling a function
invertRelationships to look up all relationships and create entries
representing their inverses. Uniqueness is enforced by calling
dedupCacheData to detect duplicate entries and merge their attributes
and relationships.

Both invertRelationships and dedupCacheData are fairly expensive as they
involve traversing the entire list of relationships (which are stored on
CacheData as a generic Collection) and building new objects based on
these. These functions also need parse keys as they encounter them,
leading to each key being parsed multiple times.

It would be more efficient to store the entries and relationships in
a data structure that guarantees uniqueness and bidrectionality of
relationships (and can do so efficiently).  Create a class
KubernetesCacheData that stores cache items as we are processing them
and that supports adding items or (bidirectional) relationships between
items.  Update the caching logic to add items to this structure instead
of creating individual CacheData entries, and then ask the structure
to give us a full list of CacheData entries at the end.