refactor(kubernetes): Remove duplicated code in cache conversion
For each manifest we encounter, we call both convertAsResource
to create an infrastructure cache key as well as convertAsArtifact
to create a logical cache key. There is fair amount of duplication
between the two methods.

Make the following changes:
(1) Move convertAsArtifact to be a called from convertAsResource
and remove the duplicated code
(2) Remove the code to add artifact relationships from
annotatedRelationships and instead add the necessary relationship
right after calling convertAsArtifact (if that call produced an
artifact).  This avoids needing to generate the artifact twice
(once for the actual cache entry and once for the relationship)
and also keeps the artifact code in one place.