perf(kubernetes): Only create applications for some kinds
After generating all of the cache data keys, we currently call
getClusterRelationships to populate application-cluster
relationships. This call ends up creating an application for
every kubernetes manifest instead of only those that have a
cluster relationship.

Instead of calling getClusterRelationships after generating the
CacheData entries, update the code that creates the manifest-cluster
and manifest-application relationships to also create an
application-cluster relationship.

In addition to increasing performance of caching itself, this change
should also greatly improve the performance of some cache lookups
(such as the search endpoint) as it will greatly reduce the number
of applications in the cache.