refactor(kubernetes): Process pod metrics outside of parallel loop
We currently fetch pod metrics in a parallel stream because most
of the time is spent in IO waiting for kubectl to return the metrics.
We're also processing the metrics in parallel, which blocks moving
to the new KubernetesCacheData object as it is not threadsafe.

Rather than make it threadsafe (which would add some overhead and
complexity) just move the processing of the received metrics out
of the parallel stream and into a sequential stream. We'll still
fetch the metrics in a parallel stream, it's just the relatively
fast post-processing that will now be sequential.