perf(kubernetes): Replace stratify function
After generating a Collection<CacheData> of kubernetes cache data,
we always then group this data by the CacheData's group and perform
some filtering of the data.

The grouping and filtering requires parsing the cache key again, so
it would be more efficient if we did this as part of KubernetesCacheData
before writing out the CacheData entries.