perf(kubernetes): Improve performance of API group and version
These two classes each use a static variable to hold all instances
of the class that are ever created. Before creating a new instance
of the class, they check to see if there is already a matching
instance and if so return that instance instead of creating a
new one.

While this might be a reasonable pattern for objects that are expensive
to create, here the cost of keeping this object pool is extremely high
vs. the cost of just creating an object each time we need it.

Partly this is because we don't use an efficient data structure for
storing exiting objects, and need to loop over an array of all existing
objects each time we want to fetch/create a new one.  This is compounded
by the fact that that loop is in a synchronized block for thread safety.

While we could make this object pool more efficient by using a better
data structure and handling synchronization better, as noted above these
objects are so cheap to create that it's not worth the effort.

Also add tests to the class as part of this refactor.