refactor(kubernetes): Change from == to equals for KubernetesKind
Most comparisons of KubernetesKind use == instead of equals(), which
works because we currently keep a static pool of all KubernetesKinds
and never create duplicates.

The cost of keeping this static pool is much higher than the benefit,
so move to using .equals() instead of == to prepare for a future where
there can be two different KubernetesKind objects that are equal.