refactor(k8s): defer to computedOmitKinds
previously, we were modifying omitKinds when an account was initialized.
this caused calls to `equals` to fail because they didn't match
configured omitKinds. this change defers to `omitKindsComputed` so that
checking new and existing credentials is accurate according to how the
account it configured. this allows us to only initialize modified or new
accounts rather than every credential.