refactor(kubernetes): Create KubernetesKindRegistry
The KubernetesKind class is responsible for both providing information
about a given KubernetesKind and for keeping a registry of all
registered kinds.  The registry of all kinds is stored as a list which
is very inefficient to search.

In preparation for making this logic more efficient, create a
KubernetesKindRegistry that supports registering kind or getting
registered kinds. The implementation will be the exact same
(inefficient) one that is currently directly in KubernetesKind but
we have now defined an API and can write tests against it, and in
a future commit replace it with a more efficient implementation.