perf(kubernetes): Use null instead of optional
We create a lot of ScopedKind objects to look up kinds, so the
cost of creating an Optional each time is non-negligible.

In general using Optional can be more readable than null but in this
case we'd rather the performance of null (and it's just as easy to read
here).