perf(kubernetes): Filter unclassified kinds from search query
When performing searches, we attempt to interpret the input
type as a Spinnaker kind and to search the corresponding
Kubernetes kinds. For example, if a user inputs serverGroups,
we'll translate that to add replicaSets and daemonSets to the
types to search.

If the input type is *not* a Spinnaker kind, it gets mapped to
KubernetesSpinnakerKindMap.SpinnakerKind.UNCLASSIFIED. Currently
this is causing the search logic to search *every* Kubernetes
kind that is also unclassified which is expensive and not what
we want.

If the input string is not a Spinnaker kind, just skip adding any
extra kubernetes kinds to search.