perf(kubernetes): Improve performance of kubernetes search
Instead of processing all relationships of applications/clusters
that match, and later filtering to those that are of types of
interest, push the filtering logic down. Only fetch relationships
that point to a kind of interest.

We can also simplify the post-filtering logic; the only time the
result's type and group will differ is when the group is a
KubernetesKind and the type is the corresponding SpinnakerKind.
But if we're searching for a SpinnakerKind, the typesToSearch
will already have all mapped KubernetesKind's so it's redundant
to also check if the SpinnakerKind is in typesToMatch.

This also removes the need to separately keep track of typesToSearch
and typesToMatch.