fix(appengine): Fix threading bugs in AppengineMutexRepository
There are a couple of threading bugs here:
(1) We do an unsyncronized containsKey + put to generate the mutex
for a given String, which can lead to two threads holding different
Mutexes for the same String.
(2) The Mutex itself has a bug whereby two threads calling lock can
actually acquire different semaphores.

Remove the Mutex class entirely and replace it with a ReentrantLock
which supports the same operations. Use a ConcurrentHashMap and the
atomic computeIfAbsent to store the mappings from keys to locks.