perf(redis): Replace String.format with String.join (#3836) (#3915)
When caching large Kubernetes clusters, ~25% of allocated objects per
caching cycle are from calling String.format on cache keys.

String.format is significantly more expensive than String.join as
it parses the input as a regular expression. In this case, we can
replace the calls to .format() with calls to .join() while keeping
exactly the same functionality.