refactor(kubernetes): Convert a few more classes to Java
Also, as part of this conversion update
KubernetesAtomicOperationConverterHelper to use generics, which
avoids a lot of necessary casting and provides extra type safety.
This extra type safety surfaced a bug in the resize server group
converter where we were incorrectly converting the description to a
deploy manifest description which compiled before but no longer does.