refactor(kubernetes): Remove registered field from KubernetesKind
The registered field is only ever used in in the
KubernetesResourcePropertyRegistry, where we short-circuit if a
kind isn't registered. But the definition of a registered kind is
one that has been registered with KubernetesResourcePropertyRegistry,
so there's no need to also store this on the kind itself.

Remove this field, and update the property registry to figure out
for itself if a kind is registered.