refactor(kubernetes): Add global and account-specific registry
We currently have a single KubernetesResourcePropertyRegistry that
tracks both global properties and has a map of account-specific
properties.

Rather than store the account-specific properties in a global object,
let's create a new class to store account-specific properties,
AccountResourcePropertyRegistry. The credentials for an account will
contain an AccountResourcePropertyRegistry, making it easy to look
up relevant information from code that just has the credentials.

The global properties will live in a GlobalResourcePropertyRegistry;
each AccountResourcePropertyRegistry will contain a reference to a
singleton GlobalResourcePropertyRegistry and will fall back to that
registry if needed.

Finally, to simplify the few cases where we are looking up
account-specific properties without credentials, create a class
KubernetesResourcePropertyResolver that handles the boilerplate of
resolving accounts and casting them to Kubernetes accounts.

Finally, replace all uses of the global singleton property registry with
the one on the account relevant to the operation.