refactor(gce): Add a way to run batches of paginated requests
This isn't necessarily the most efficient because we wait for ALL the
page(n) results to come in before sending off a batch request for
the page(n+1) results. I suppose you could request just the first pages
in a batch and request all subsequent pages in unbatched requests,
starting them immediately upon receipt of the previous page. But I think
this is probably good enough.