fix(kubernetes): Use an account-specific kind registry (#3985)
* fix(kubernetes): Use an account-specific kind registry

The primary goal of the changes here is to split up the global
KubernetesKindRegistry (primarily accessed through the static
functions on KubernetesKind) into a separate kind registries for
each account.

Before this change, KubernetesKind was a KubernetesKind.ScopedKind
(name + api group) as well as some properties (ex: namespaced,
hasClusterRelationship). The first change this commit makes is
to rename these classes:
* KubernetesKind is now what KubernetesKind.ScopedKind was before,
just a pointer to a name and api group (with no properties or
other data)
* KubernetesKindProperties is now what KubernetsKind was before;
it has a KubernetesKind as well as additional properties

The advantage of this renaming is that the vast majority of times
we are using a KubernetesKind in the codebase, we actually don't
need the properties so these usages aren't affected at all by
removing the properties from KubernetesKind.

Then, we've added a KubernetesKindRegistry onto each account object
that allows for registering mappings from kinds to their properties
as well as fetching the properties for a given kind. The
account-specific registry falls back to a global registry of built-in
kinds if we look up something that's not in the registry.

Any of the places that were using KubernetesKind before and that did
actually need the properties were then updated to look up the properties
using the registry on the appropriate account (sometimes needing to
thread it down a call or two).

Finally, prior to this change the old implementation of a kind registry
handled lookig up things by alias (so if you asked for "rs" it would
actually return the properties registered for "replicaSet"). This logic
was moved out of the registry and into the new KubernetesKind class;
now if you try to create a KubernetesKind from the string "rs" it will
just give you back a "replicaSet", which means the logic can be removed
from the registry. (We can do this because the alias functionality is
only supported for a few built-in kinds and is not supported for crds.)

This is more change than I normally like to include in a single commit,
but given the intertwined nature of these refactors, my attempts to
split this up actually led to a more complicated diff, so hopefully
this is readable enough with the above description.

* style(kubernetes): Address code review comments

Replace wildcard imports and remove unused variable