fix(kubernetes): Use consistent casing for kind names
Kind names are currently serialized preserving their case; prior
to the recent refactor, we had a global registry of kinds that
effectively enforced a single casing for every kind. Now that this
registry is gone, there are some places where we are referring to
a replicaset as ReplicaSet and others as replicaSet which is causing
issues.

In order to resolve this without resorting to adding back a global
registry, convert kind names to lower case before serializing (and
thus before generating cache keys).