fix(kubernetes): Revert consistent casing for kind
The prior commit to always serialize kinds as lowercase
is still something we should do longer-term to reduce the
fragility of code that is depending on particular casings
(ex: replicaSet).

That being said, it turns out there are more places depending
on the current casing than I initially thought. In particular,
the logic to substitute parts of a kubernetes manifest (ex:
to replace a config map with the fully-qualified config map
with a version) depends on the kind being camelCased.

As we're approaching 1.16, I think that this is a more invasive
change that I'd like to do right now, and would also like to
decouple that change from the more important work to refactor
Kubernetes kinds.  So this commit rolls back the change that
always converts kinds to lower-case; the following commit
fixes the issue with the refactor a different (and less
invasive way).