fix(kubernetes): Use consistent casing for built-in kinds
A lot of code depends on case-sensitive comparisons between
kinds; the recent refactor removed the global kind registry
so when Spinnaker encounters ReplicaSet it is leaving it cased
as ReplicaSet instead of (as before) converting it to replicaSet.

This change builds a small collection of the built-in kinds and
handles that conversion to Spinnaker-canonical casing when we're
reading kinds. This is only done for built-in kinds as these are
the only ones where we have built-in logic that depends on the case.

A later PR (after 1.16) will re-implement always converting to lower
case and will fix places where we're incorrectly doing case-sensitive
comparisons.