feat(artifacts/gitRepo): add support for git repo
adds support for a git repo artifact type which will clone a given git
repo, package it up as a tar.gz and return the value. callers of this
artifact type should expect a .tar.gz and act accordingly, decompressing
if necessary. this impl. doesn't include support for cloning the
repository at a particular history depth as jgit doesn't support that
option. the alternative is to perform git operations by shelling out to
git but i opted not to do that because it makes authentication much more
difficult than it is _with_ jgit.