refactor(kubernetes): Change lists to immutable sets
We store a list of kinds and kinds to omit on each account;
these lists are never modified after they are created, and
we only ever do set-like operations on them (contains, size).

In order to make lookups more efficient, and prevent bugs (and
make the following refactor easier), change the lists to immutable
sets.