perf(kubernetes): Defer checking valid kinds until after startup
Instead of blocking the startup of clouddriver while we check
whether we can read every kind in every account, defer these
checks until first use.

Encapsulate this logic in an inner class called KindValidator;
this class handles keeping track of what kinds are valid for an
account as well as whether metrics are readable for the account.
It will only ever attempt to read a kind once (the first time it
is asked), and short-circuits just like the earlier logic based on
the contents of kinds and omitKinds.