perf(kubernetes): Defer loading namespaces until first use
The kubernetes v2 caching agents store a list of namespaces of
interest, and have a complex system of calling loadNamespaces()
to refresh this list then getNamespaces() to fetch them.

It turns out that we always just call loadNamespaces() immediately
before a call to getNamespaces() so the cache is doing nothing.
Remove the cached namespaces variable and all calls to
loadNamespaces(). Instead just have getNamespaces() directly look
up the namespaces (which was happening anyway before).

The main benefit is that now we can remove the call to loadNamespaces()
to prime the cache from the caching agent constructor. This means that
caching agents don't need to communicate with the cluster in their
constructor, so startup will not be blocked waiting for calls to
an unreachable cluster to time out.