refactor(gce): Prepare zonal caching agent for an abstract superclass
About  90% of the code from GoogleZonalServerGroupCachingAgent was just
copied into GoogleRegionalServerGroupCachingAgent. I'm going to extract
all the common code into an abstract base class.

Pretty much all that will be left after I do that are the methods
marked `// @Override`. Those methods will be the abstract ones in the
superclass.