fix(gce): fix a small bug in GoogleRegionalServerGroupCachingAgent
This bug is likely pretty innocuous since callers are looking for a
specific key amongst the pendingOnDemandRequests and returning an extra
one won't hurt anything, but I might as well fix it.

It looks like this bug likely exists in other caching agents, too, but
since it's very long-standing and the impact is negligible, I'm going to
ignore that.