perf(kubernetes): Defer checking if metrics are enabled
Rather than checking if metrics are enabled before scheduling
caching agents, push the responsibility for determining this
down to the metrics caching agent. This means we'll always
schedule the metrics caching agent, but it will be a no-op
if metrics are disabled.

The advantage here is that we don't need to make a live lookup
to determine if metrics are readable during startup.