refactor(kubernetes): Add Nonnull annotation to handler kinds
In an upcoming commit, we're going to be storing the KubernetesKind
and SpinnakerKind associated with handlers in an ImmutableMap, which
won't allow null values.

As we already have null objects for each of these (KubernetesKind.NONE
and SpinnakerKind.UNCLASSIFIED) there's no need to ever have these
methods return null.  Add annotations to the interface and
implementations, and update the one implementation returning null
to instead return SpinnakerKind.UNCLASSIFIED.