refactor(kubernetes): Let KubernetesSpinnakerKindMap initialize itself
It's a bit strange that GlobalResourcePropertyRegistry is responsible for
initializing KubernetesSpinnakerKindMap. (It takes the singleton kind map
in the constructor, mutates it, then forgets about it.) This is a vestige
of the way the code worked before the kind registries were created.

Instead, let KubernetesSpinnakerKindMap handle its own initialization; its
constructor should expect the same list of KubernetesHandlers and should
register the kinds found on each of them.

This also means that we can make the register method private as it's only
called from the constructor. If you were to guess that the next commit would
remove the register method entirely and make the class immutable, you would
probably be correct...