refactor(kubernetes): Enforce that defaultProperties are not null
We've added the NonNull annotation to GlobalResourcePropertyRegistry.get
because we fall back to the properties for KubernetesKind.NONE for a
kind that's not in the registry. While it is statically true that there
will be properties for the NONE kind (because
KubernetesUnregisteredCustomResourceHandler handles the NONE kind) it's
a bit confusing to see how this defaulting happens and is liable to break
if one day we remove the NONE kind from that handler.

Instead, let's be more explicit and expect to receive the unregistered custom
resource handler bean in the constructor of the global registry, and set that
as the default to return. Then if we somehow don't have this default value,
autowiring will fail at startup (instead of causing null-pointer exceptions
later).

This makes it much more clear that get() always returns a non-null value,
even though it always did.