fix(kubernetes): Fix deserialization of account property (#4022)
The change in spinnaker/clouddriver#4017 broke deserialization of
the account property. This is due to some confusing inheritence
where KubernetesAtomicOperationDescription has a field 'account'
that overrides the default method in the CredentialsNameable
interface. Adding @JsonProperty("credentials") to the interface
method broke deserialization.

This can be fixed by adding an explicit @JsonProperty("account")
to the field in KubernetesAtomicOperationDescription. As this is
subtle, also add a test to validate this (and prevent it from
breaking in the future). While it might be better to rename the
account field, it's used all over the V1 provider in groovy code,
so this change is much safer as it keeps the interface of the
class to consumers the same.

The test that is added passes before spinnaker/clouddriver#4017,
breaks after that change, and passes again with this change.