refactor(kubernetes): Split CredentialFactory to V1 and V2
A lot of the unecessary casts that we do when creating credentials
stem from the fact that we have a single CredentialFactory with
different methods for creating V1 and V2 credentials.

Instead, make a generic interface KubernetesCredentialFactory that
pulls out the common logic to default methods.  (These default methods
should probably both eventually live somewhere else, but I think
having them on the interface is a resonable solution for now.)

Then give both the V1 and V2 credentials an inner factory class that
implements that interface and allows us to create credentials in a
type-safe way. It also logically makes more sense, as the factory for
credentials lives with the credentials it's creating.