refactor(kubernetes): Push getSpinnakerKindMap to V1/V2
The getSpinnakerKindMap function on the base
KubernetesNamedAccountCredentials class is needed to support
sending the kind map to the UI. It forks using an if block
depending on which implementation we're using, which means
we should push the logic down to the actual implementations.

One solution would have been to keep the kind map at the top
level and just implement an "add custom resources" in the
V2 provider class (that is a no-op in V1). I didn't do that
because:
(1) It's strange that there's logic in this class anyway, which
is just supposed to hold some metadata and delegate down to
the actual credentials.
(2) Ultimately, it probably makes sense for the V2 provider to
implement its own account-specific KubernetesSpinnakerKindMap
instead of building this map on the fly. Pushing this down to
the V1/V2 implementations allows us to do this later.