refactor(kubernetes): Split some shared and V2-only code
KubernetesSpinnakerKindMap is only used by the V2 provider, but
SpinnakerKind is needed by both. In order to support having the
V1 provider use SpinnakerKind but not the KubernetesSpinnakerKindMap,
move SpinnakerKind to the top level.

Instead of building all the Kubernetes beans in KubernetesConfiguration,
move the construction of V2-only beans to KubernetesV2Configuration.